+++ Node.Js/bp/backup-2026-02-10T14-34-00.052Z/saludos.js
@@ -1,0 +1,8 @@
+
+
+function saludar(nombre) {
+    return `Bienvenido, ${nombre}!`;
+}
+
+  
+module.exports.saludar = saludar;