+++ Node.Js/bp/backup-2026-02-10T14-35-00.025Z/saludos.js
@@ -1,0 +1,8 @@
+
+
+function saludar(nombre) {
+    return `Bienvenido, ${nombre} ha iniciado correctamente!`;
+}
+
+  
+module.exports.saludar = saludar;