+++ Node.Js/intro.js
@@ -1,0 +1,3 @@
+//globalThis: variable global dentro de la app
+//windows: espacio del pc de navegador
+//global: nuestro entorno nodejs